+++ src/Core/SignalRApp.Application/Services/CategoryService.cs
@@ -1,4 +1,10 @@
+using System.Net;
+using AutoMapper;
 using SignalRApp.Application.Constants;
+using SignalRApp.Application.Features.Category.Commands.Create;
+using SignalRApp.Application.Features.Category.Commands.Delete;
+using SignalRApp.Application.Features.Category.Commands.Update;
+using SignalRApp.Application.Helpers;
 using SignalRApp.Application.Interfaces.Repository;
 using SignalRApp.Application.Interfaces.Service;
 using SignalRApp.Application.Interfaces.UnitOfWork;
@@ -6,7 +12,7 @@ using SignalRApp.Domain.Entities;
 
 namespace SignalRApp.Application.Services;
 
-public class CategoryService(ICategoryRepository categoryRepository,IUnitOfWork unitOfWork): ICategoryService
+public class CategoryService(ICategoryRepository categoryRepository,IUnitOfWork unitOfWork,IMapper mapper): ICategoryService
 {
     public async Task<Category> GetByIdAsync(int id)
     {
@@ -24,21 +30,39 @@ public class CategoryService(ICategoryRepository categoryRepository,IUnitOfWork 
         return await categoryRepository.GetAllAsync();
     }
 
-    public async ValueTask AddAsync(Category category, CancellationToken cancellationToken)
+    public async Task<ServiceResult> AddAsync(CreateCategoryCommandRequest request, CancellationToken cancellationToken)
     {
+        var category = mapper.Map<Category>(request);
         await categoryRepository.AddAsync(category);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask UpdateAsync(Category category, CancellationToken cancellationToken)
+    public async Task<ServiceResult> UpdateAsync(UpdateCategoryCommandRequest request, CancellationToken cancellationToken)
     {
-        categoryRepository.Update(category);
+        var category = await categoryRepository.GetByIdAsync(request.Id);
+        if (category is null)
+        {
+            return ServiceResult.Failure(CategoryConstant.NotFound);
+        }
+
+        var updatedCategory = mapper.Map(request, category);
+        categoryRepository.Update(updatedCategory);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask DeleteAsync(Category category, CancellationToken cancellationToken)
-    {
+    public async Task<ServiceResult> DeleteAsync(DeleteCategoryCommandRequest request, CancellationToken cancellationToken)
+    {  
+        var category = await categoryRepository.GetByIdAsync(request.Id);
+        if (category is null)
+        {
+            return ServiceResult.Failure(CategoryConstant.NotFound);
+        }
+        
         categoryRepository.Delete(category);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 }
