+++ src/Core/SignalRApp.Application/Services/ContactService.cs
@@ -1,4 +1,10 @@
+using System.Net;
+using AutoMapper;
 using SignalRApp.Application.Constants;
+using SignalRApp.Application.Features.Contact.Commands.Create;
+using SignalRApp.Application.Features.Contact.Commands.Delete;
+using SignalRApp.Application.Features.Contact.Commands.Update;
+using SignalRApp.Application.Helpers;
 using SignalRApp.Application.Interfaces.Repository;
 using SignalRApp.Application.Interfaces.Service;
 using SignalRApp.Application.Interfaces.UnitOfWork;
@@ -6,7 +12,7 @@ using SignalRApp.Domain.Entities;
 
 namespace SignalRApp.Application.Services;
 
-public class ContactService(IContactRepository contactRepository,IUnitOfWork unitOfWork): IContactService
+public class ContactService(IContactRepository contactRepository,IUnitOfWork unitOfWork,IMapper mapper): IContactService
 {
     public async Task<Contact> GetByIdAsync(int id)
     {
@@ -24,21 +30,38 @@ public class ContactService(IContactRepository contactRepository,IUnitOfWork uni
         return await contactRepository.GetAllAsync();
     }
 
-    public async ValueTask AddAsync(Contact contact, CancellationToken cancellationToken)
+    public async Task<ServiceResult> AddAsync(CreateContactCommandRequest request, CancellationToken cancellationToken)
     {
+        var contact = mapper.Map<Contact>(request);
         await contactRepository.AddAsync(contact);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask UpdateAsync(Contact contact, CancellationToken cancellationToken)
+    public async Task<ServiceResult> UpdateAsync(UpdateContactCommandRequest request, CancellationToken cancellationToken)
     {
-        contactRepository.Update(contact);
+        var contact = await contactRepository.GetByIdAsync(request.Id);
+        if (contact is null)
+        {
+            return ServiceResult.Failure(ContactConstant.NotFound);
+        }
+
+        var updatedContact = mapper.Map(request, contact);
+        contactRepository.Update(updatedContact);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask DeleteAsync(Contact contact, CancellationToken cancellationToken)
+    public async Task<ServiceResult> DeleteAsync(DeleteContactCommandRequest request, CancellationToken cancellationToken)
     {
+        var contact = await contactRepository.GetByIdAsync(request.Id);
+        if (contact is null)
+        {
+            return ServiceResult.Failure(ContactConstant.NotFound);
+        }
+        
         contactRepository.Delete(contact);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 }
