+++ src/Core/SignalRApp.Application/Services/FeatureService.cs
@@ -1,4 +1,10 @@
+using System.Net;
+using AutoMapper;
 using SignalRApp.Application.Constants;
+using SignalRApp.Application.Features.Feature.Commands.Create;
+using SignalRApp.Application.Features.Feature.Commands.Delete;
+using SignalRApp.Application.Features.Feature.Commands.Update;
+using SignalRApp.Application.Helpers;
 using SignalRApp.Application.Interfaces.Repository;
 using SignalRApp.Application.Interfaces.Service;
 using SignalRApp.Application.Interfaces.UnitOfWork;
@@ -6,7 +12,7 @@ using SignalRApp.Domain.Entities;
 
 namespace SignalRApp.Application.Services;
 
-public class FeatureService(IFeatureRepository featureRepository,IUnitOfWork unitOfWork): IFeatureService
+public class FeatureService(IFeatureRepository featureRepository,IUnitOfWork unitOfWork,IMapper mapper): IFeatureService
 {
     public async Task<Feature> GetByIdAsync(int id)
     {
@@ -24,21 +30,38 @@ public class FeatureService(IFeatureRepository featureRepository,IUnitOfWork uni
         return await featureRepository.GetAllAsync();
     }
 
-    public async ValueTask AddAsync(Feature feature, CancellationToken cancellationToken)
+    public async Task<ServiceResult> AddAsync(CreateFeatureCommandRequest request, CancellationToken cancellationToken)
     {
+        var feature = mapper.Map<Feature>(request);
         await featureRepository.AddAsync(feature);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask UpdateAsync(Feature feature, CancellationToken cancellationToken)
+    public async Task<ServiceResult> UpdateAsync(UpdateFeatureCommandRequest request, CancellationToken cancellationToken)
     {
-        featureRepository.Update(feature);
+        var feature = await featureRepository.GetByIdAsync(request.Id);
+        if (feature is null)
+        {
+            return ServiceResult.Failure(FeatureConstant.NotFound);
+        }
+
+        var updatedFeature = mapper.Map(request, feature);
+        featureRepository.Update(updatedFeature);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask DeleteAsync(Feature feature, CancellationToken cancellationToken)
+    public async Task<ServiceResult> DeleteAsync(DeleteFeatureCommandRequest request, CancellationToken cancellationToken)
     {
+        var feature = await featureRepository.GetByIdAsync(request.Id);
+        if (feature is null)
+        {
+            return ServiceResult.Failure(FeatureConstant.NotFound);
+        }
+        
         featureRepository.Delete(feature);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 }
