+++ src/Core/SignalRApp.Application/Services/TestimonialService.cs
@@ -1,4 +1,10 @@
+using System.Net;
+using AutoMapper;
 using SignalRApp.Application.Constants;
+using SignalRApp.Application.Features.Testimonial.Commands.Create;
+using SignalRApp.Application.Features.Testimonial.Commands.Delete;
+using SignalRApp.Application.Features.Testimonial.Commands.Update;
+using SignalRApp.Application.Helpers;
 using SignalRApp.Application.Interfaces.Repository;
 using SignalRApp.Application.Interfaces.Service;
 using SignalRApp.Application.Interfaces.UnitOfWork;
@@ -6,7 +12,7 @@ using SignalRApp.Domain.Entities;
 
 namespace SignalRApp.Application.Services;
 
-public class TestimonialService(ITestimonialRepository testimonialRepository,IUnitOfWork unitOfWork): ITestimonialService
+public class TestimonialService(ITestimonialRepository testimonialRepository,IUnitOfWork unitOfWork,IMapper mapper): ITestimonialService
 {
     public async Task<Testimonial> GetByIdAsync(int id)
     {
@@ -24,21 +30,37 @@ public class TestimonialService(ITestimonialRepository testimonialRepository,IUn
         return await testimonialRepository.GetAllAsync();
     }
 
-    public async ValueTask AddAsync(Testimonial testimonial, CancellationToken cancellationToken)
+    public async Task<ServiceResult> AddAsync(CreateTestimonialCommandRequest request, CancellationToken cancellationToken)
     {
+        var testimonial = mapper.Map<Testimonial>(request);
         await testimonialRepository.AddAsync(testimonial);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask UpdateAsync(Testimonial testimonial, CancellationToken cancellationToken)
+    public async Task<ServiceResult> UpdateAsync(UpdateTestimonialCommandRequest request, CancellationToken cancellationToken)
     {
-        testimonialRepository.Update(testimonial);
+        var testimonial = await testimonialRepository.GetByIdAsync(request.Id);
+        if (testimonial is null)
+        {
+            return ServiceResult.Failure(TestimonialConstant.NotFound);
+        }
+
+        var updatedTestimonial = mapper.Map(request, testimonial);
+        testimonialRepository.Update(updatedTestimonial);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 
-    public async ValueTask DeleteAsync(Testimonial testimonial, CancellationToken cancellationToken)
+    public async Task<ServiceResult> DeleteAsync(DeleteTestimonialCommandRequest request, CancellationToken cancellationToken)
     {
+        var testimonial = await testimonialRepository.GetByIdAsync(request.Id);
+        if (testimonial is null)
+        {
+            return ServiceResult.Failure(TestimonialConstant.NotFound);
+        }
         testimonialRepository.Delete(testimonial);
         await unitOfWork.SaveChangesAsync(cancellationToken);
+        return ServiceResult.Success(HttpStatusCode.NoContent);
     }
 }
